solid split between sync and async
