+++ tracing-modality/src/common/options.rs
